Rework some of the non-instance-dependent methods to be static.
Methods that will be used as filter callbacks and never change based on the state of the object are really not appropriate as non-static methods, especially when we need to be able to hook and un-hook them from filter callbacks. For example, get_postmeta_mapping() will always return the same, hard-coded array, as nothing within the data store class was altering the (now-removed) $postmeta_mapping property.

Having a static reference makes it much easier to add/remove actions and filters, since there's only a single instance to worry about.