Permit multisite tests to fail (for now), as there's a bug within the WooCommerce core test suite that prevents it from passing.
When running WooCommerce's core test suite (with or without the custom orders table plugin) with WP_MULTISITE=1, we get the following error:

1) WC_Tests_Setup_Functions::test_wizard_in_cart_payment_gateways
Failed asserting that two arrays are equal.
--- Expected
+++ Actual
@@ @@
 Array (
-    'paypal' => false
+    'stripe' => true
+    'ppec_paypal' => true

/home/travis/build/liquidweb/woocommerce-order-tables/vendor/woocommerce/woocommerce/tests/unit-tests/setup/functions.php:38

The multisite test should remain in the test matrix as a reminder, but a failure for code we don't control shouldn't break the build.