Use --prefer-lowest when installing Composer dependencies for PHP 7.0
Something odd seems to be preventing Travis from completing when PHPUnit 6.5.14 is installed via Composer by PHP 7.2 (the config.platform value for this project's composer.json file) and run under PHP 7.0, yet the same tests work fine when the PHPUnit 6.5.14 PHAR file is installed and executed using PHP 7.0.

Working theory is that *another* dependency is seeing that a more modern PHP is supported and, as a result, it's getting loaded in. Based on the changes when I re-install dependencies with config.platform set to 7.0 on WooCommerce, it appears that it's due to doctrine/instantiator, which raised its minimum PHP version to 7.1 in version 1.1.0: https://github.com/doctrine/instantiator/releases/tag/1.1.0.

Related to https://github.com/woocommerce/woocommerce/pull/24114