Dynamically rewrite the PHPUnit "core" test suite at runtime
This is some weird, wild stuff, my friends:

The plugin traditionally has two test suites: "plugin" (e.g. just the plugin code), and "core" (running the WooCommerce core test suite with the WooCommerce Custom Orders Table enabled).

The thinking here is "anything that works in WooCommerce without the plugin active should still work once it's turned on."

Since the phpunit.xml(.dist) file doesn't support variables, we have to hijack it with tests/bootstrap.php: once we've determined which version of WooCommerce we're testing against (default is "master"), the bootstrap file will grab the `PHPUnit\Util\Configuration` instance associated with `phpunit.xml.dist` in this repo, find the definition for the `<testsuite name="core" />`, then replace instances of "%WC_VERSION%" with the suffix corresponding to the local copy of WooCommerce (e.g. "3.8" in "woocommerce/woocommerce-src-3.8").

As a bonus, we can now exclude breaking core tests on a per-version basis, which is pretty nifty.