Since we can use features newer than PHP 5.2, start modernizing the codebase
The main thing here is splitting some of the shared functionality between the order + refund data stores into a trait, so we can maintain it in a single place.

We're also moving towards more modern coding standards (Per https://github.com/woocommerce/woocommerce-example-package, WooCommerce packages should use PSR-4 autoloading), so this is the first introduction of namespaces into the project (with more to come in the future).