Introduce a new QueryIterator class
This commit defines a new `QueryIterator` class, which is based on the `WP_CLI\Iterators\Query` class. Our iterator also extends the original, adding support for the `Countable` interface, along with better inline documentation and tests. Please note that there are not yet any *uses* of the new iterator in the codebase.

There are a few reasons for importing this into the project:

1. Leverage iterators rather than ad-hoc `while()` loops, with automatic support for adjusting offsets and creating corresponding `COUNT(*)` queries
2. Ensure that we have the iterator available regardless of whether or not WP-CLI is present on the target system. This will be important when it comes to #66

This commit duplicates some code from #159, and should be considered a blocker for that PR. It also is the first use of PSR-4 autoloading in the repository, which will eventually be the norm (see #161).